add python script to implement better serial comms
Create file with tables of command keys and codes.  This will be the first implementation of purely binary serial communication between a PC and the TF Node.  The goal is to move away from sending commands as Strings and instead use a pure binary bitstream with a common table of command-codes between the two devices.